requirements.txt
google-genai
langchain==1.2.6
